typescript_app: clean up fabric registry
