+++ utility-emissions-channel/typescript_app/src/blockchain-gateway/I-gateway.ts
